# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fratelanza Management Console API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects with optional filters
      parameters:
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /projects/{id}/payment:
    post:
      operationId: logPayment
      tags: [projects]
      summary: Log a payment received for a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Updated project after payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
  /projects/{id}/team:
    get:
      operationId: getProjectTeam
      tags: [projects]
      summary: Get team members for a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: addTeamMember
      tags: [projects]
      summary: Add a team member to a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Added team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /projects/{id}/team/{memberId}:
    delete:
      operationId: removeTeamMember
      tags: [projects]
      summary: Remove a team member from a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: memberId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Freelancers ───────────────────────────────────────────────────────
  /freelancers:
    get:
      operationId: listFreelancers
      tags: [freelancers]
      summary: List freelancers with optional filters
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: spec
          schema:
            type: string
      responses:
        "200":
          description: Freelancers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Freelancer"
    post:
      operationId: createFreelancer
      tags: [freelancers]
      summary: Add a new freelancer/instructor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreelancerInput"
      responses:
        "201":
          description: Created freelancer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Freelancer"
...
  /freelancers/{code}:
    patch:
      operationId: updateFreelancer
      tags: [freelancers]
      summary: Update a freelancer
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreelancerUpdate"
      responses:
        "200":
          description: Updated freelancer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Freelancer"
        "404":
          description: Not found
    delete:
      operationId: deleteFreelancer
      tags: [freelancers]
      summary: Delete a freelancer
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Clients ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, type, projectName, clientPrice, totalCost, netProfit, status, date]
      properties:
        id:
          type: integer
        type:
          type: string
        projectName:
          type: string
        clientName:
          type: ["string", "null"]
        clientPrice:
          type: number
        totalCost:
          type: number
        netProfit:
          type: number
        freelancerName:
          type: ["string", "null"]
        freelancerCommission:
          type: number
        startDate:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        status:
          type: string
        paidAmount:
          type: number
        remainingAmount:
          type: number
        nextPaymentDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        date:
          type: string
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            team:
              type: array
              items:
                $ref: "#/components/schemas/TeamMember"
    ProjectInput:
      type: object
      required: [type, projectName, clientPrice]
      properties:
        type:
          type: string
        projectName:
          type: string
        clientName:
          type: string
        clientPrice:
          type: number
        totalCost:
          type: number
        netProfit:
          type: number
        freelancerName:
          type: string
        freelancerCommission:
          type: number
        startDate:
          type: string
        deadline:
          type: string
        status:
          type: string
        paidAmount:
          type: number
        remainingAmount:
          type: number
        nextPaymentDate:
          type: string
        notes:
          type: string
        team:
          type: array
          items:
            $ref: "#/components/schemas/TeamMemberInput"
    ProjectUpdate:
      type: object
      properties:
        projectName:
          type: string
        clientName:
          type: string
        clientPrice:
          type: number
        totalCost:
          type: number
        netProfit:
          type: number
        freelancerCommission:
          type: number
        startDate:
          type: string
        deadline:
          type: string
        status:
          type: string
        paidAmount:
          type: number
        remainingAmount:
          type: number
        nextPaymentDate:
          type: string
        notes:
          type: string
        team:
          type: array
          items:
            $ref: "#/components/schemas/TeamMemberInput"
    PaymentInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        nextPaymentDate:
          type: string
    TeamMember:
      type: object
      required: [id, projectId, freelancerName, commission]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        freelancerName:
          type: string
        commission:
          type: number
    TeamMemberInput:
      type: object
      required: [freelancerName, commission]
      properties:
        freelancerName:
          type: string
        commission:
          type: number

    # ── Freelancer ─────────────────────────────────────────────────────
    Freelancer:
      type: object
      required: [code, name, earned, balance, rating]
      properties:
        code:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        spec:
          type: ["string", "null"]
        position:
          type: ["string", "null"]
        earned:
          type: number
        balance:
          type: number
        rating:
          type: number
    FreelancerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        spec:
          type: string
        position:
          type: string
        earned:
          type: number
        balance:
          type: number
    FreelancerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        spec:
          type: string
        position:
          type: string
        earned:
          type: number
        balance:
          type: number
        rating:
          type: number

    # ── Client ─────────────────────────────────────────────────────────
...